refacto: add an abstract PointCloud system
Now, the Potree system is based on this new PointCloud system. It can
now be used as a basis for other type of point clouds, like Entwine
Point Tile.

BREAKING CHANGE: The signature of the `PotreeLayer` constructor has
changed.
