+++ src/Core/PotreeNode.js
@@ -1,83 +1,67 @@
 import * as THREE from 'three';
+import PointCloudNode from 'Core/PointCloudNode';
 
 // Create an A(xis)A(ligned)B(ounding)B(ox) for the child `childIndex` of one aabb.
 // (PotreeConverter protocol builds implicit octree hierarchy by applying the same
 // subdivision algo recursively)
 const dHalfLength = new THREE.Vector3();
 
-class PotreeNode {
+class PotreeNode extends PointCloudNode {
     constructor(numPoints = 0, childrenBitField = 0, layer) {
-        this.numPoints = numPoints;
+        super(numPoints, layer);
         this.childrenBitField = childrenBitField;
-        this.children = [];
-        this.layer = layer;
-        this.name = '';
-        this.bbox = new THREE.Box3();
-        this.sse = -1;
+        this.id = '';
+        this.depth = 0;
         this.baseurl = layer.source.baseurl;
     }
 
     add(node, indexChild, root) {
-        this.children.push(node);
-        node.parent = this;
-        node.name = this.name + indexChild;
-        this.createChildAABB(indexChild, node.bbox);
-        if ((node.name.length % this.layer.hierarchyStepSize) == 0) {
-            node.baseurl = `${root.baseurl}/${node.name.substr(root.name.length)}`;
+        super.add(node, indexChild);
+        node.id = this.id + indexChild;
+        node.depth = node.id.length;
+        if ((node.id.length % this.layer.hierarchyStepSize) == 0) {
+            node.baseurl = `${root.baseurl}/${node.id.substr(root.id.length)}`;
         } else {
             node.baseurl = root.baseurl;
         }
     }
 
-    createChildAABB(childIndex, box) {
+    createChildAABB(node, childIndex) {
         // Code inspired from potree
-        box.copy(this.bbox);
-        this.bbox.getCenter(box.max);
-        dHalfLength.copy(box.max).sub(this.bbox.min);
+        node.bbox.copy(this.bbox);
+        this.bbox.getCenter(node.bbox.max);
+        dHalfLength.copy(node.bbox.max).sub(this.bbox.min);
 
         if (childIndex === 1) {
-            box.min.z += dHalfLength.z;
-            box.max.z += dHalfLength.z;
+            node.bbox.min.z += dHalfLength.z;
+            node.bbox.max.z += dHalfLength.z;
         } else if (childIndex === 3) {
-            box.min.z += dHalfLength.z;
-            box.max.z += dHalfLength.z;
-            box.min.y += dHalfLength.y;
-            box.max.y += dHalfLength.y;
+            node.bbox.min.z += dHalfLength.z;
+            node.bbox.max.z += dHalfLength.z;
+            node.bbox.min.y += dHalfLength.y;
+            node.bbox.max.y += dHalfLength.y;
         } else if (childIndex === 0) {
             //
         } else if (childIndex === 2) {
-            box.min.y += dHalfLength.y;
-            box.max.y += dHalfLength.y;
+            node.bbox.min.y += dHalfLength.y;
+            node.bbox.max.y += dHalfLength.y;
         } else if (childIndex === 5) {
-            box.min.z += dHalfLength.z;
-            box.max.z += dHalfLength.z;
-            box.min.x += dHalfLength.x;
-            box.max.x += dHalfLength.x;
+            node.bbox.min.z += dHalfLength.z;
+            node.bbox.max.z += dHalfLength.z;
+            node.bbox.min.x += dHalfLength.x;
+            node.bbox.max.x += dHalfLength.x;
         } else if (childIndex === 7) {
-            box.min.add(dHalfLength);
-            box.max.add(dHalfLength);
+            node.bbox.min.add(dHalfLength);
+            node.bbox.max.add(dHalfLength);
         } else if (childIndex === 4) {
-            box.min.x += dHalfLength.x;
-            box.max.x += dHalfLength.x;
+            node.bbox.min.x += dHalfLength.x;
+            node.bbox.max.x += dHalfLength.x;
         } else if (childIndex === 6) {
-            box.min.y += dHalfLength.y;
-            box.max.y += dHalfLength.y;
-            box.min.x += dHalfLength.x;
-            box.max.x += dHalfLength.x;
+            node.bbox.min.y += dHalfLength.y;
+            node.bbox.max.y += dHalfLength.y;
+            node.bbox.min.x += dHalfLength.x;
+            node.bbox.max.x += dHalfLength.x;
         }
-    }
-
-    getChildByName(name) {
-        if (this.name === name) {
-            return this;
-        }
-        const charIndex = this.name.length;
-        for (const child of this.children) {
-            if (child.name[charIndex] == name[charIndex]) {
-                return child.getChildByName(name);
-            }
-        }
-        throw new Error(`Cannot find node with name '${name}'`);
     }
 
     get octreeIsLoaded() {
@@ -85,20 +69,11 @@ class PotreeNode {
     }
 
     get url() {
-        return `${this.baseurl}/r${this.name}.${this.layer.source.extension}`;
-    }
-
-    load() {
-        // Query octree/HRC if we don't have children potreeNode yet.
-        if (!this.octreeIsLoaded) {
-            this.loadOctree();
-        }
-
-        return this.layer.source.fetcher(this.url, this.layer.source.networkOptions).then(this.layer.source.parse);
+        return `${this.baseurl}/r${this.id}.${this.layer.source.extension}`;
     }
 
     loadOctree() {
-        const octreeUrl = `${this.baseurl}/r${this.name}.${this.layer.source.extensionOctree}`;
+        const octreeUrl = `${this.baseurl}/r${this.id}.${this.layer.source.extensionOctree}`;
         return this.layer.source.fetcher(octreeUrl, this.layer.source.networkOptions).then((blob) => {
             const view = new DataView(blob);
             const stack = [];
