+++ src/Layer/PointCloudLayer.js
@@ -1,0 +1,372 @@
+import * as THREE from 'three';
+import GeometryLayer from 'Layer/GeometryLayer';
+import PointsMaterial, { MODE } from 'Renderer/PointsMaterial';
+import Picking from 'Core/Picking';
+
+const point = new THREE.Vector3();
+const bboxMesh = new THREE.Mesh();
+const box3 = new THREE.Box3();
+bboxMesh.geometry.boundingBox = box3;
+
+function initBoundingBox(elt, layer) {
+    elt.tightbbox.getSize(box3.max);
+    box3.max.multiplyScalar(0.5);
+    box3.min.copy(box3.max).negate();
+    elt.obj.boxHelper = new THREE.BoxHelper(bboxMesh);
+    elt.obj.boxHelper.geometry = new THREE.Geometry().fromBufferGeometry(elt.obj.boxHelper.geometry.toNonIndexed());
+    elt.obj.boxHelper.computeLineDistances();
+    elt.obj.boxHelper.material = elt.childrenBitField ? new THREE.LineDashedMaterial({ dashSize: 0.25, gapSize: 0.25 }) : new THREE.LineBasicMaterial();
+    elt.obj.boxHelper.material.color.setHex(0);
+    elt.obj.boxHelper.material.linewidth = 2;
+    elt.obj.boxHelper.frustumCulled = false;
+    elt.obj.boxHelper.position.copy(elt.tightbbox.min).add(box3.max);
+    elt.obj.boxHelper.autoUpdateMatrix = false;
+    elt.obj.boxHelper.layers.mask = layer.bboxes.layers.mask;
+    layer.bboxes.add(elt.obj.boxHelper);
+    elt.obj.boxHelper.updateMatrix();
+    elt.obj.boxHelper.updateMatrixWorld();
+}
+
+function computeScreenSpaceError(context, pointSize, spacing, elt, distance) {
+    if (distance <= 0) {
+        return Infinity;
+    }
+    const pointSpacing = spacing / 2 ** elt.depth;
+    // Estimate the onscreen distance between 2 points
+    const onScreenSpacing = context.camera.preSSE * pointSpacing / distance;
+    // [  P1  ]--------------[   P2   ]
+    //     <--------------------->      = pointsSpacing (in world coordinates)
+    //                                  ~ onScreenSpacing (in pixels)
+    // <------>                         = pointSize (in pixels)
+    return Math.max(0.0, onScreenSpacing - pointSize);
+}
+
+function markForDeletion(elt) {
+    if (elt.obj) {
+        elt.obj.material.visible = false;
+        if (__DEBUG__) {
+            if (elt.obj.boxHelper) {
+                elt.obj.boxHelper.material.visible = false;
+            }
+        }
+    }
+
+    if (!elt.notVisibleSince) {
+        elt.notVisibleSince = Date.now();
+        // Set .sse to an invalid value
+        elt.sse = -1;
+    }
+    for (const child of elt.children) {
+        markForDeletion(child);
+    }
+}
+
+/**
+ * The basis for all point clouds related layers.
+ *
+ * @property {boolean} isPointCloudLayer - Used to checkout whether this layer
+ * is a PointCloudLayer. Default is `true`. You should not change this, as it is
+ * used internally for optimisation.
+ * @property {THREE.Group|THREE.Object3D} group - Contains the created
+ * `THREE.Points` meshes, usually with an instance of a `THREE.Points` per node.
+ * @property {THREE.Group|THREE.Object3D} bboxes - Contains the bounding boxes
+ * (`THREE.Box3`) of the tree, usually one per node.
+ * @property {number} octreeDepthLimit - The depth limit at which to stop
+ * browsing the octree. Can be used to limit the browsing, without having to
+ * edit manually the source of the point cloud. No limit by default (`-1`).
+ * @property {number} [pointBudget=2000000] - Maximum number of points to
+ * display at the same time. This influences the performance of rendering.
+ * Default to two millions points.
+ * @property {number} [sseThreshold=2] - Threshold of the **S**creen **S**pace
+ * **E**rror. Default to `2`.
+ * @property {number} [pointSize=4] - The size (in pixels) of the points.
+ * Default to `4`.
+ * @property {THREE.Material|PointsMaterial} [material=new PointsMaterial] - The
+ * material to use to display the points of the cloud. Be default it is a new
+ * `PointsMaterial`.
+ * @property {number} [mode=MODE.COLOR] - The displaying mode of the points.
+ * Values are specified in `PointsMaterial`.
+ */
+class PointCloudLayer extends GeometryLayer {
+    /**
+     * Constructs a new instance of point cloud layer.
+     * Constructs a new instance of a Point Cloud Layer. This should not be used
+     * directly, but rather implemented using `extends`.
+     *
+     * @constructor
+     * @extends GeometryLayer
+     *
+     * @param {string} id - The id of the layer, that should be unique. It is
+     * not mandatory, but an error will be emitted if this layer is added a
+     * {@link View} that already has a layer going by that id.
+     * @param {Object} [config] - Optional configuration, all elements in it
+     * will be merged as is in the layer. For example, if the configuration
+     * contains three elements `name, protocol, extent`, these elements will be
+     * available using `layer.name` or something else depending on the property
+     * name. See the list of properties to know which one can be specified.
+     */
+    constructor(id, config = {}) {
+        super(id, new THREE.Group(), config);
+        this.isPointCloudLayer = true;
+        this.protocol = 'pointcloud';
+
+        this.group = config.group || new THREE.Group();
+        this.object3d.add(this.group);
+        this.bboxes = config.bboxes || new THREE.Group();
+        this.bboxes.visible = false;
+        this.object3d.add(this.bboxes);
+        this.group.updateMatrixWorld();
+
+        // default config
+        this.octreeDepthLimit = config.octreeDepthLimit || -1;
+        this.pointBudget = config.pointBudget || 2000000;
+        this.pointSize = config.pointSize === 0 || !isNaN(config.pointSize) ? config.pointSize : 4;
+        this.sseThreshold = config.sseThreshold || 2;
+        this.material = config.material || {};
+        this.material = this.material.isMaterial ? config.material : new PointsMaterial(config.material);
+        this.material.defines = this.material.defines || {};
+        this.mode = config.mode || MODE.COLOR;
+    }
+
+    preUpdate(context, changeSources) {
+        // See https://cesiumjs.org/hosted-apps/massiveworlds/downloads/Ring/WorldScaleTerrainRendering.pptx
+        // slide 17
+        context.camera.preSSE =
+            context.camera.height /
+                (2 * Math.tan(THREE.MathUtils.degToRad(context.camera.camera3D.fov) * 0.5));
+
+        if (this.material) {
+            this.material.visible = this.visible;
+            this.material.opacity = this.opacity;
+            this.material.transparent = this.opacity < 1;
+            this.material.size = this.pointSize;
+        }
+
+        // lookup lowest common ancestor of changeSources
+        let commonAncestor;
+        for (const source of changeSources.values()) {
+            if (source.isCamera || source == this) {
+                // if the change is caused by a camera move, no need to bother
+                // to find common ancestor: we need to update the whole tree:
+                // some invisible tiles may now be visible
+                return [this.root];
+            }
+            if (source.obj === undefined) {
+                continue;
+            }
+            // filter sources that belong to our layer
+            if (source.obj.isPoints && source.obj.layer == this) {
+                if (!commonAncestor) {
+                    commonAncestor = source;
+                } else {
+                    commonAncestor = source.findCommonAncestor(commonAncestor);
+
+                    if (!commonAncestor) {
+                        return [this.root];
+                    }
+                }
+            }
+        }
+
+        if (commonAncestor) {
+            return [commonAncestor];
+        }
+
+        // Start updating from hierarchy root
+        return [this.root];
+    }
+
+    update(context, layer, elt) {
+        elt.visible = false;
+
+        if (this.octreeDepthLimit >= 0 && this.octreeDepthLimit < elt.depth) {
+            markForDeletion(elt);
+            return;
+        }
+
+        // pick the best bounding box
+        const bbox = (elt.tightbbox ? elt.tightbbox : elt.bbox);
+        elt.visible = context.camera.isBox3Visible(bbox, this.object3d.matrixWorld);
+        if (!elt.visible) {
+            markForDeletion(elt);
+            return;
+        }
+
+        elt.notVisibleSince = undefined;
+        point.copy(context.camera.camera3D.position).sub(this.object3d.position);
+
+        // only load geometry if this elements has points
+        if (elt.numPoints > 0) {
+            if (elt.obj) {
+                if (elt.obj.material.update) {
+                    elt.obj.material.update(this.material);
+                } else {
+                    elt.obj.material.copy(this.material);
+                }
+                if (__DEBUG__) {
+                    if (this.bboxes.visible) {
+                        if (!elt.obj.boxHelper) {
+                            initBoundingBox(elt, layer);
+                        }
+                        elt.obj.boxHelper.visible = true;
+                        elt.obj.boxHelper.material.color.r = 1 - elt.sse;
+                        elt.obj.boxHelper.material.color.g = elt.sse;
+                    }
+                }
+            } else if (!elt.promise) {
+                const distance = Math.max(0.001, bbox.distanceToPoint(point));
+                // Increase priority of nearest node
+                const priority = computeScreenSpaceError(context, layer.pointSize, layer.spacing, elt, distance) / distance;
+                elt.promise = context.scheduler.execute({
+                    layer,
+                    requester: elt,
+                    view: context.view,
+                    priority,
+                    redraw: true,
+                    earlyDropFunction: cmd => !cmd.requester.visible || !this.visible,
+                }).then((pts) => {
+                    if (this.onPointsCreated) {
+                        this.onPointsCreated(layer, pts);
+                    }
+
+                    elt.obj = pts;
+                    // store tightbbox to avoid ping-pong (bbox = larger => visible, tight => invisible)
+                    elt.tightbbox = pts.tightbbox;
+
+                    // make sure to add it here, otherwise it might never
+                    // be added nor cleaned
+                    this.group.add(elt.obj);
+                    elt.obj.updateMatrixWorld(true);
+
+                    elt.promise = null;
+                }, (err) => {
+                    if (err.isCancelledCommandException) {
+                        elt.promise = null;
+                    }
+                });
+            }
+        }
+
+        if (elt.children && elt.children.length) {
+            const distance = bbox.distanceToPoint(point);
+            elt.sse = computeScreenSpaceError(context, layer.pointSize, layer.spacing, elt, distance) / this.sseThreshold;
+            if (elt.sse >= 1) {
+                return elt.children;
+            } else {
+                for (const child of elt.children) {
+                    markForDeletion(child);
+                }
+            }
+        }
+    }
+
+    postUpdate() {
+        this.displayedCount = 0;
+        for (const pts of this.group.children) {
+            if (pts.material.visible) {
+                const count = pts.geometry.attributes.position.count;
+                pts.geometry.setDrawRange(0, count);
+                this.displayedCount += count;
+            }
+        }
+
+        if (this.displayedCount > this.pointBudget) {
+            // 2 different point count limit implementation, depending on the potree source
+            if (this.supportsProgressiveDisplay) {
+                // In this format, points are evenly distributed within a node,
+                // so we can draw a percentage of each node and still get a correct
+                // representation
+                const reduction = this.pointBudget / this.displayedCount;
+                for (const pts of this.group.children) {
+                    if (pts.material.visible) {
+                        const count = Math.floor(pts.geometry.drawRange.count * reduction);
+                        if (count > 0) {
+                            pts.geometry.setDrawRange(0, count);
+                        } else {
+                            pts.material.visible = false;
+                        }
+                    }
+                }
+                this.displayedCount *= reduction;
+            } else {
+                // This format doesn't require points to be evenly distributed, so
+                // we're going to sort the nodes by "importance" (= on screen size)
+                // and display only the first N nodes
+                this.group.children.sort((p1, p2) => p2.userData.node.sse - p1.userData.node.sse);
+
+                let limitHit = false;
+                this.displayedCount = 0;
+                for (const pts of this.group.children) {
+                    const count = pts.geometry.attributes.position.count;
+                    if (limitHit || (this.displayedCount + count) > this.pointBudget) {
+                        pts.material.visible = false;
+                        limitHit = true;
+                    } else {
+                        this.displayedCount += count;
+                    }
+                }
+            }
+        }
+
+        const now = Date.now();
+        for (let i = this.group.children.length - 1; i >= 0; i--) {
+            const obj = this.group.children[i];
+            if (!obj.material.visible && (now - obj.userData.node.notVisibleSince) > 10000) {
+                // remove from group
+                this.group.children.splice(i, 1);
+
+                if (Array.isArray(obj.material)) {
+                    for (const material of obj.material) {
+                        material.dispose();
+                    }
+                } else {
+                    obj.material.dispose();
+                }
+                obj.geometry.dispose();
+                obj.material = null;
+                obj.geometry = null;
+                obj.userData.node.obj = null;
+
+                if (__DEBUG__) {
+                    if (obj.boxHelper) {
+                        obj.boxHelper.removeMe = true;
+                        if (Array.isArray(obj.boxHelper.material)) {
+                            for (const material of obj.boxHelper.material) {
+                                material.dispose();
+                            }
+                        } else {
+                            obj.boxHelper.material.dispose();
+                        }
+                        obj.boxHelper.geometry.dispose();
+                    }
+                }
+            }
+        }
+
+        if (__DEBUG__) {
+            this.bboxes.children = this.bboxes.children.filter(b => !b.removeMe);
+        }
+    }
+
+    pickObjectsAt(view, mouse, radius, target = []) {
+        return Picking.pickPointsAt(view, mouse, radius, this, target);
+    }
+
+    getObjectToUpdateForAttachedLayers(meta) {
+        if (meta.obj) {
+            const p = meta.parent;
+            if (p && p.obj) {
+                return {
+                    element: meta.obj,
+                    parent: p.obj,
+                };
+            } else {
+                return {
+                    element: meta.obj,
+                };
+            }
+        }
+    }
+}
+
+export default PointCloudLayer;
