+++ test/unit/potreelayerprocessing.js
@@ -1,54 +1,84 @@
 import assert from 'assert';
 import PotreeLayer from 'Layer/PotreeLayer';
+import PotreeNode from 'Core/PotreeNode';
 
 describe('preUpdate PotreeLayer', function () {
     const context = { camera: { height: 1, camera3D: { fov: 1 } } };
+    const layer = {
+        id: 'a',
+        source: { baseurl: 'server.geo' },
+        hierarchyStepSize: 1,
+    };
+    layer.root = new PotreeNode(4000, 0, layer);
+    layer.root.bbox.setFromArray([1000, 1000, 1000, 0, 0, 0]);
+
+    layer.root.add(new PotreeNode(3000, 0, layer), 1, layer.root);
+    layer.root.children[0].obj = { layer, isPoints: true };
+    layer.root.add(new PotreeNode(3000, 0, layer), 2, layer.root);
+    layer.root.children[1].obj = { layer, isPoints: true };
+    layer.root.add(new PotreeNode(3000, 0, layer), 3, layer.root);
+    layer.root.children[2].obj = { layer, isPoints: true };
+
+    layer.root.children[0].add(new PotreeNode(2000, 0, layer), 1, layer.root);
+    layer.root.children[0].children[0].obj = { layer, isPoints: true };
+    layer.root.children[0].add(new PotreeNode(2000, 0, layer), 2, layer.root);
+    layer.root.children[0].children[1].obj = { layer, isPoints: true };
+    layer.root.children[1].add(new PotreeNode(2000, 0, layer), 1, layer.root);
+    layer.root.children[1].children[0].obj = { layer, isPoints: true };
+    layer.root.children[2].add(new PotreeNode(2000, 0, layer), 2, layer.root);
+    layer.root.children[2].children[0].obj = { layer, isPoints: true };
+    layer.root.children[2].add(new PotreeNode(2000, 0, layer), 3, layer.root);
+    layer.root.children[2].children[1].obj = { layer, isPoints: true };
+
+    layer.root.children[0].children[0].add(new PotreeNode(1000, 0, layer), 1, layer.root);
+    layer.root.children[0].children[0].children[0].obj = { layer, isPoints: true };
+    layer.root.children[0].children[0].add(new PotreeNode(1000, 0, layer), 5, layer.root);
+    layer.root.children[0].children[0].children[1].obj = { layer, isPoints: true };
+    layer.root.children[0].children[1].add(new PotreeNode(1000, 0, layer), 4, layer.root);
+    layer.root.children[0].children[1].children[0].obj = { layer, isPoints: true };
+    layer.root.children[2].children[1].add(new PotreeNode(1000, 0, layer), 1, layer.root);
+    layer.root.children[2].children[1].children[0].obj = { layer, isPoints: true };
+    layer.root.children[2].children[1].add(new PotreeNode(1000, 0, layer), 2, layer.root);
+    layer.root.children[2].children[1].children[1].obj = { layer, isPoints: true };
+    layer.root.children[2].children[1].add(new PotreeNode(1000, 0, layer), 3, layer.root);
+    layer.root.children[2].children[1].children[2].obj = { layer, isPoints: true };
+    layer.root.children[2].children[1].add(new PotreeNode(1000, 0, layer), 4, layer.root);
+    layer.root.children[2].children[1].children[3].obj = { layer, isPoints: true };
 
     it('should return root if no change source', () => {
-        const layer = { root: {} };
         const sources = new Set();
-        assert.equal(
+        assert.deepStrictEqual(
             layer.root,
             PotreeLayer.prototype.preUpdate.call(layer, context, sources)[0]);
     });
 
     it('should return root if no common ancestors', () => {
-        const layer = { root: {}, id: 'a' };
-        const elt1 = { name: '12', obj: { layer: 'a', isPoints: true } };
-        const elt2 = { name: '345', obj: { layer: 'a', isPoints: true } };
         const sources = new Set();
-        sources.add(elt1);
-        sources.add(elt2);
-        assert.equal(
+        sources.add(layer.root.children[0].children[0]);
+        sources.add(layer.root.children[2].children[1]);
+        assert.deepStrictEqual(
             layer.root,
             PotreeLayer.prototype.preUpdate.call(layer, context, sources)[0]);
     });
 
     it('should return common ancestor', () => {
-        const layer = { root: {}, id: 'a' };
-        const elt1 = { name: '123', obj: { layer: 'a', isPoints: true } };
-        const elt2 = { name: '12567', obj: { layer: 'a', isPoints: true } };
-        const elt3 = { name: '122', obj: { layer: 'a', isPoints: true } };
         const sources = new Set();
-        sources.add(elt1);
-        sources.add(elt2);
-        sources.add(elt3);
-        layer.root.getChildByName = (name) => {
-            assert.equal('12', name);
-        };
-        PotreeLayer.prototype.preUpdate.call(layer, context, sources);
+        sources.add(layer.root.children[2].children[0]);
+        sources.add(layer.root.children[2].children[1]);
+        sources.add(layer.root.children[2].children[1].children[2]);
+        sources.add(layer.root.children[2].children[1].children[3]);
+        assert.deepStrictEqual(
+            layer.root.children[2],
+            PotreeLayer.prototype.preUpdate.call(layer, context, sources)[0]);
     });
 
     it('should not search ancestors if layer are different root if no common ancestors', () => {
-        const layer = { root: {}, id: 'a' };
-        const elt1 = { name: '12', obj: { layer: 'a', isPoints: true } };
-        const elt2 = { name: '13', obj: { layer: 'b', isPoints: true } };
         const sources = new Set();
-        sources.add(elt1);
-        sources.add(elt2);
-        layer.root.getChildByName = (name) => {
-            assert.equal('12', name);
-        };
-        PotreeLayer.prototype.preUpdate.call(layer, context, sources);
+        sources.add(layer.root.children[2].children[0]);
+        sources.add(layer.root.children[2].children[1].children[3]);
+        layer.root.children[2].children[1].children[3].obj = { layer: {}, isPoints: true };
+        assert.deepStrictEqual(
+            layer.root.children[2].children[0],
+            PotreeLayer.prototype.preUpdate.call(layer, context, sources)[0]);
     });
 });
